Rename user folder into userCard and move it to shared ui. Move postItem folder to shared ui. Update imports accordingly
