+++ src/features/comments/ui/BaseCommentItem.tsx
@@ -1,14 +1,13 @@
 'use client'
 
-import { ReactNode } from 'react'
 import clsx from 'clsx'
+import { ReactNode } from 'react'
 
-import { MetaInfo } from './MetaInfo'
-import { Button } from '@/shared/ui'
+import { Button, UserCard } from '@/shared/ui'
 import { CommentLikeButton } from './CommentLikeButton'
+import { MetaInfo } from './MetaInfo'
 
 import s from './BaseCommentItem.module.scss'
-import { UserCard } from '@/shared/user'
 
 type BaseCommentItemProps = {
   id: number
