+++ src/shared/ui/index.ts
@@ -21,6 +21,7 @@ export { ArrowIcon } from './icons'
 export { Loader } from './loader'
 export { Modal } from './modal'
 export { Pagination } from './pagination'
+export { PostItem } from './postItem'
 export { PublicationDescription } from './publicationDescription'
 export { RadioGroup, RadioGroupItem, RadioGroupRoot } from './radioGroup'
 export { Recaptcha } from './reCaptcha'
@@ -32,3 +33,4 @@ export { TextArea } from './textArea'
 export { TextField } from './textField'
 export { Typography } from './typography'
 export { ToastSnackbar } from './toastSnackbar'
+export { UserCard } from './userCard'
