+++ src/shared/ui/postItem/PostItem.module.scss
