+++ src/shared/ui/postItem/index.ts
