+++ src/shared/ui/userCard/index.ts
@@ -1,1 +1,2 @@
 export * from './ui/userCard/UserCard'
+export * from './types/userCard.types'
