+++ src/shared/ui/userCard/types/userCard.types.ts
