+++ src/shared/ui/userCard/ui/userCard/UserCard.module.scss
